Update README.md
In the role of a data analyst who frequently conducts data extractions from a larger database, the project emulates a common process in which extracted data is harnessed for in-depth analysis. Each data extraction serves various organizational needs, ranging from the creation of executive summaries to the provision of comprehensive, detailed datasets. This project aims to replicate and streamline this process, enabling the development of a repeatable system for data export, import, and analysis. It caters to the diverse needs of stakeholders.